refactor: extract useCalendarSync hook
Moves taskCalendarUrl, taskCalendarAuth, syncRetentionDays, completedTaskUids,
pendingImportFile, showImportModal, importColor, calSyncStatus, calSyncLastSynced
state plus the taskCalendarAuth persist effect into a dedicated useCalendarSync hook.
